+++ controllers/userController.ts
@@ -1,7 +1,9 @@
 import { SortOrder } from "mongoose"
 import { userModel } from "../models/userModel"
-import { Request, Response } from "express"
+import { NextFunction, Request, Response } from "express"
 import jwt, { JwtPayload } from "jsonwebtoken"
+import { IGetUserAuthInfoRequest } from "../types"
+import { ErrorHandler } from "../utils/errorHandling"
 
 const generateToken = (id: any, role: string) => {
     //'any' because imported ObjectId is not assignable to needed Object id
@@ -13,7 +15,7 @@ const generateToken = (id: any, role: string) => {
     return jwt.sign(payload, "SUPER_SECRET_KEY", { expiresIn: "3d" })
 }
 
-const createUser = async (req: Request, res: Response) => {
+const createUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
         let { name, password, email } = req.body
         let role: string
@@ -24,29 +26,24 @@ const createUser = async (req: Request, res: Response) => {
         }
 
         if (!name || !password || !email) {
-            res.status(400).json("All fields are required")
-            return
+            throw new ErrorHandler(400, "All fields are required")
         }
         let user = await userModel.findOne({ name })
         if (user) {
-            res.status(400).json("This name is already associated with an account") //reason for 'any'
-            return
+            throw new ErrorHandler(400, "This name is already associated with an account") //reason for 'any'
         }
         user = await userModel.findOne({ email })
         if (user) {
-            res.status(400).json("This email is already associated with an account") //reason for 'any'
-            return
+            throw new ErrorHandler(400, "This email is already associated with an account") //reason for 'any'
         }
 
         let passwordRegExp = /^(?=.*[0-9])[a-zA-Z0-9!@#$%^&*]{8,1024}$/
         let emailRegExp = /^(([^<>()[\]\.,;:\s@\"]+(\.[^<>()[\]\.,;:\s@\"]+)*)|(\".+\"))@(([^<>()[\]\.,;:\s@\"]+\.)+[^<>()[\]\.,;:\s@\"]{2,})$/i
         if (!passwordRegExp.test(password)) {
-            res.status(400).json("Invalid password")
-            return
+            throw new ErrorHandler(400, "Invalid password")
         }
         if (!emailRegExp.test(email)) {
-            res.status(400).json("Invalid email")
-            return
+            throw new ErrorHandler(400, "Invalid email")
         }
 
         user = await new userModel({ name, password, email, role }).save()
@@ -55,100 +52,100 @@ const createUser = async (req: Request, res: Response) => {
 
         res.status(200).json({ _id: user._id, name: user.name, email: user.email, token })
     } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
+        next(error)
     }
 }
 
-const authoriseUser /*N/I*/ = async (req: Request, res: Response | any) => {
+const authoriseUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        let { passwordOrEmail, password } = req.body
-        let user = await userModel.findOne({ passwordOrEmail })
+        let { nameOrEmail, password } = req.body
+        let user = await userModel.findOne({ $or: [{ name: nameOrEmail }, { email: nameOrEmail }] })
 
         if (!user) {
-            res.status(400).json("Invalid name, email or password")
-            return
+            throw new ErrorHandler(400, "Invalid name, email or password")
         }
 
         const isValidPassword = password === user.password
         if (!isValidPassword) {
-            res.status(400).json("Invalid name, email or password")
-            return
+            throw new ErrorHandler(400, "Invalid name, email or password")
         }
 
         const token = generateToken(user._id, user.role)
 
         res.status(200).json({ _id: user._id, name: user.name, email: user.email, token })
     } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
+        next(error)
     }
 }
 
-const updateUser = async (req: Request, res: Response | any) => {
+const updateUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
+        let authorizationError = (req as IGetUserAuthInfoRequest).error
+
+        if (authorizationError) {
+            throw new ErrorHandler(authorizationError.statusCode, authorizationError.message)
+        }
+
         let userId = req.params.userId
         let { name, password, email } = req.body
 
         if (!name || !password || !email) {
-            res.status(400).json("All fields are required")
-            return
+            throw new ErrorHandler(400, "All fields are required")
         }
 
         let user = await userModel.findOne({ name })
         if (user && user._id.toString() !== userId) {
-            res.status(400).json("This name is already associated with an account") //reason for 'any'
-            return
+            throw new ErrorHandler(400, "This name is already associated with an account")
         }
         user = await userModel.findOne({ email })
         if (user && user._id.toString() !== userId) {
-            res.status(400).json("This email is already associated with an account") //reason for 'any'
-            return
+            throw new ErrorHandler(400, "This email is already associated with an account")
         }
 
         let passwordRegExp = /^(?=.*[0-9])[a-zA-Z0-9!@#$%^&*]{8,1024}$/
         let emailRegExp = /^(([^<>()[\]\.,;:\s@\"]+(\.[^<>()[\]\.,;:\s@\"]+)*)|(\".+\"))@(([^<>()[\]\.,;:\s@\"]+\.)+[^<>()[\]\.,;:\s@\"]{2,})$/i
         if (!passwordRegExp.test(password)) {
-            res.status(400).json("Invalid password")
-            return
+            throw new ErrorHandler(400, "Invalid password")
         }
         if (!emailRegExp.test(email)) {
-            res.status(400).json("Invalid email")
-            return
+            throw new ErrorHandler(400, "Invalid email")
         }
 
         user = await userModel.findByIdAndUpdate(userId, { name, password, email }, { new: true })
         res.status(200).json({ name, password, email, id: user!._id })
     } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
+        next(error)
     }
 }
 
-const deleteUser = async (req: Request, res: Response) => {
+const deleteUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
+        let authorizationError = (req as IGetUserAuthInfoRequest).error
+
+        if (authorizationError) {
+            throw new ErrorHandler(authorizationError.statusCode, authorizationError.message)
+        }
+
         let { userId } = req.params
         let user = await userModel.findByIdAndDelete(userId).select("name email")
         res.status(200).json(user)
     } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
+        next(error)
     }
 }
 
-const findUser = async (req: Request, res: Response) => {
+const findUser = async (req: Request, res: Response, next: NextFunction) => {
     try {
         let userId = req.params.userId
 
         let user = await userModel.findById(userId).select("name email")
         res.status(200).json(user)
     } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
+        next(error)
     }
 }
 
-const getUsers = async (req: Request, res: Response | any) => {
+const getUsers = async (req: Request, res: Response, next: NextFunction) => {
     try {
         let { page, name, order } = req.query
         let sortOrder: SortOrder
@@ -156,8 +153,7 @@ const getUsers = async (req: Request, res: Response | any) => {
         if (!order) {
             order = "1"
         } else if (order !== "-1" && order !== "1") {
-            res.status(400).json("Invalid order")
-            return
+            throw new ErrorHandler(400, "Invalid order")
         }
 
         order === "1" ? (sortOrder = 1) : (sortOrder = -1)
@@ -181,8 +177,7 @@ const getUsers = async (req: Request, res: Response | any) => {
             .skip(Number(page) * usersPerPage)
         res.status(200).json(users)
     } catch (error) {
-        console.log(error)
-        res.status(500).json(error)
+        next(error)
     }
 }
